manage-conversion/requirements.txt

Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.15
Werkzeug==2.0.1
psycopg2-binary
flask_migrate
flask
flask_cors
aiven-client
python-dotenv
eventlet
pika
google-cloud-storage
py-healthcheck